+++ omigami/__init__.py
@@ -12,7 +12,6 @@ from ._version import get_versions
 
 __version__ = get_versions()["version"]
 del get_versions
-from .spectra_matching import MS2DeepScore, Spec2Vec
 from omigami.plotting import MoleculePlotter
 
 ROOT_DIR = Path(__file__).parent.parent
