+++ omigami/spectra_matching/__init__.py
@@ -1,0 +1,2 @@
+from omigami.spectra_matching.spec2vec import Spec2Vec
+from omigami.spectra_matching.ms2deepscore import MS2DeepScore
